+++ src/components/sensCalcDisplay/single/SensCalcDisplaySingle.tsx
@@ -11,6 +11,7 @@ const TOTAL_SENSE = 'TotalSensitivity';
 const BEAM_SIZE = 'SynthBeamSize';
 const VALUE = 'value';
 const UNITS = 'units';
+const SHOW_ERROR_ON_LINE = false; // Leave this as I expect this will return
 
 interface SensCalcDisplaySingleProps {
   sensCalc: any;
@@ -46,16 +47,16 @@ export default function SensCalcDisplaySingle({ sensCalc, show }: SensCalcDispla
           <Grid item xs={2}>
             <IconButton
               style={{ cursor: 'hand' }}
-              onClick={sensCalc?.status === STATUS_OK ? IconClicked : null}
+              onClick={sensCalc?.statusGUI === STATUS_OK ? IconClicked : null}
             >
               <StatusIcon
                 ariaTitle={t('sensitivityCalculatorResults.status', {
-                  status: t('statusLoading.' + sensCalc?.status),
+                  status: t('statusLoading.' + sensCalc?.statusGUI),
                   error: sensCalc?.error
                 })}
                 testId="statusId"
                 icon
-                level={sensCalc?.status}
+                level={sensCalc?.statusGUI}
                 size={SIZE}
               />
             </IconButton>
@@ -72,7 +73,7 @@ export default function SensCalcDisplaySingle({ sensCalc, show }: SensCalcDispla
           )}
           {hasError() && (
             <Grid item xs={10}>
-              {sensCalc?.error}
+              {SHOW_ERROR_ON_LINE && sensCalc?.error}
             </Grid>
           )}
         </Grid>