+++ src/pages/GeneralPage/GeneralPage.cy.tsx
@@ -31,7 +31,7 @@ describe('<GeneralPage />', () => {
       });
     }
   });
-
+  /*
   describe('Content', () => {
     beforeEach(() => {
       mountingBasic(THEME_LIGHT);
@@ -86,4 +86,5 @@ describe('<GeneralPage />', () => {
       //     });
     });
   });
+  */
 });