+++ src/pages/ObservationPage/ObservationPage.tsx
@@ -16,6 +16,7 @@ import Observation from '../../utils/types/observation';
 import { Proposal } from '../../utils/types/proposal';
 import { validateObservationPage } from '../../utils/proposalValidation';
 import {
+  BANDWIDTH_TELESCOPE,
   PATH,
   STATUS_ERROR,
   STATUS_INITIAL,
@@ -75,7 +76,7 @@ export default function ObservationPage() {
     let result = STATUS_INITIAL;
     filteredByObservation(obs.id)?.forEach(rec => {
       if (typeof rec !== 'undefined') {
-        switch (rec.status) {
+        switch (rec.statusGUI) {
           case STATUS_ERROR:
             result = STATUS_ERROR;
             return;
@@ -162,7 +163,7 @@ export default function ObservationPage() {
 
   const setSensCalcForTargetGrid = (observationId: string, target: Target, sensCalc: any) => {
     const tmpTO = [{ targetId: target.id, observationId: observationId, sensCalc: sensCalc }];
-    elementsS.forEach(rec => {
+    elementsS.forEach((rec: TargetObservation) => {
       if (
         (rec: { targetId: number; observationId: string }) =>
           rec.targetId !== target.id || rec.observationId !== observationId
@@ -171,7 +172,7 @@ export default function ObservationPage() {
       }
     });
     setElementsS(tmpTO);
-    if (sensCalc?.status === STATUS_PARTIAL) {
+    if (sensCalc?.statusGUI === STATUS_PARTIAL) {
       getSensCalcData(target);
     }
   };
@@ -218,19 +219,18 @@ export default function ObservationPage() {
   };
 
   const addObservationTarget = (target: Target) => {
-    const ss = {
-      id: target.id,
-      title: target.name,
-      statusGUI: STATUS_PARTIAL,
-      error: ''
-    };
     const rec: TargetObservation = {
       observationId: currObs.id,
       targetId: target.id,
-      sensCalc: ss
+      sensCalc: {
+        id: target.id,
+        title: target.name,
+        statusGUI: STATUS_PARTIAL,
+        error: ''
+      }
     };
     addTargetObservationStorage(rec);
-    setSensCalcForTargetGrid(currObs.id, target, ss);
+    setSensCalcForTargetGrid(currObs.id, target, rec.sensCalc);
   };
 
   function filterRecords(id: number) {
@@ -317,11 +317,11 @@ export default function ObservationPage() {
       },
       {
         field: 'telescope',
-        headerName: t('arrayConfiguration.short'),
-        flex: 0.5,
+        headerName: t('observingBand.label'),
+        flex: 1.5,
         disableClickEventBubbling: true,
-        renderCell: (e: { row: { telescope: number } }) =>
-          t(`arrayConfiguration.${e.row.telescope}`)
+        renderCell: (e: { row: { rec: { observingBand: string | number } } }) =>
+          BANDWIDTH_TELESCOPE[e.row.rec.observingBand]?.label
       },
       {
         field: 'subarray',