+++ src/services/axios/sensitivityCalculator/sensCalHelpers.ts
@@ -150,14 +150,9 @@ const sensCalHelpers = {
       return bandwidthValue * unitMap[bandwidthUnits];
     },
     convertBandwidthToHz(bandwidthValue, bandwidthUnits): number {
-      const unitMap: { [key: string]: number } = {
-        GHz: 1000000000,
-        MHz: 1000000,
-        KHz: 1000,
-        Hz: 1
-      };
-      if (!unitMap[bandwidthUnits]) {
-        throw new Error('Invalid bandwidth unit');
+      const unitMap = [0, 1000000000, 1000000, 1000, 1]; // Spacer, GHz, MHz, KHz, Hz
+      if (bandwidthUnits < 1 || bandwidthUnits > unitMap.length) {
+        throw new Error(`Invalid bandwidth unit: ${bandwidthUnits}`);
       }
       return bandwidthValue * unitMap[bandwidthUnits];
     }
@@ -170,7 +165,7 @@ const sensCalHelpers = {
   map: {
     getFrequencyAndBandwidthUnits(unitsField: number, telescope: number): string {
       const array = OBSERVATION.array.find(item => item.value === telescope);
-      let units = array.CentralFrequencyAndBandWidthUnits.find(item => item.value === unitsField)
+      let units = array.centralFrequencyAndBandWidthUnits.find(item => item.value === unitsField)
         ?.label;
       return units;
     }